+++ src/prime_rl/trainer/sft/train.py
@@ -119,7 +119,9 @@ def train(config: SFTTrainerConfig):
 
     # Set up the optimizer
     logger.info(f"Initializing optimizer ({config.optim})")
-    optimizer = setup_optimizer(config.optim, list(model.named_parameters()), parallel_dims)
+    optimizer = setup_optimizer(
+        config.optim, list(model.named_parameters()), parallel_dims, cpu_offload=config.model.optim_cpu_offload
+    )
 
     # Set up the learning rate scheduler
     scheduler_steps = (